--- src/Interface/SubjectwiseDetails.ts
@@ -1,11 +1,0 @@
-interface SubjectwiseDetails {
-  subjectName: string;
-  setNumber: number;
-  totalQuestionsCount: number;
-  createdBy?: string;
-  updatedBy?: string;
-  createdDate?: string;
-  updatedDate?: string;
-}
-
-export default SubjectwiseDetails;
