--- src/components/SubjectExpertDataList/SubjectList.style.scss
@@ -1,18 +1,0 @@
-.no-quizset-box {
-  text-align: center;
-  margin: 2rem;
-}
-
-.subjectlist-box {
-  margin-left: 5%;
-  margin-right: 5%;
-  .table-heading {
-    display: flex;
-    justify-content: flex-start;
-
-    .heading {
-      margin-left: 30%;
-      font-size: 1.5rem;
-    }
-  }
-}
