+++ src/screens/subjectExpert/SubjectExpert.style.scss
@@ -1,29 +1,20 @@
-.Dialog-Container {
-  display: flex;
-  flex-direction: column !important;
-  justify-content: space-around;
-  margin: 2rem;
+// .Dialog-Container {
+//   display: flex;
+//   flex-direction: column !important;
+//   justify-content: space-around;
+//   margin: 2rem;
 
-  .items {
-    margin-top: 1rem;
-  }
-  .file {
-    margin-top: 1.5rem;
-  }
-  .button {
-    margin: 1rem;
-  }
-  .error {
-    color: red;
-    font-size: 0.8rem;
-  }
-}
-
-.question-upload-buttons {
-  display: flex;
-  justify-content: space-evenly;
-  margin: 2rem;
-  width: 50%;
-  margin-left: auto;
-  margin-right: auto;
-}
+//   .items {
+//     margin-top: 1rem;
+//   }
+//   .file {
+//     margin-top: 1.5rem;
+//   }
+//   .button {
+//     margin: 1rem;
+//   }
+//   .error {
+//     color: red;
+//     font-size: 0.8rem;
+//   }
+// }
